% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mars.R
\name{anova.mars}
\alias{anova.mars}
\title{ANOVA for MARS Model Fits}
\usage{
\method{anova}{mars}(mars)
}
\arguments{
\item{mars}{a mars object}
}
\description{
ANOVA for MARS Model Fits
}
\details{
Implementation of ANOVA for mars object
}
\examples{
## Analyzing Ozone data with mars function
m <-mars(Ozone~Wind+Temp+Month,data=airquality,control=mars.control())
anova(m)
}
\seealso{
Other methods: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
