% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{h}
\alias{h}
\title{Step Function}
\usage{
h(x, s, t)
}
\arguments{
\item{x}{the data value being compared to the split point}

\item{s}{the side of the split point}

\item{t}{the split point}
}
\value{
if s=+1, this returns max(0,x-t); if s=-1, this return max(0,t-x)
}
\description{
Implementation of the step function defined on page 11 of the Friedman paper
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, 19(1), 1–67. https://doi.org/10.1214/aos/1176347963
Lin, W. (2023). Lecture Material.
}
