% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{Initialize B matrix}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{the number of points in each basis function}

\item{Mmax}{the maxium number of basis functions}
}
\value{
a data frame with an intercept column filled with ones and a column filled with
NA for each basis function.
}
\description{
Initialize a matrix for all the basis functions
}
\references{
Lin, W. (2023). Lecture Material.
}
