% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{make_B}
\alias{make_B}
\title{Make B matrix}
\usage{
make_B(model, Bfuncs)
}
\arguments{
\item{model}{a model to create basis functions for}

\item{Bfuncs}{a list of all basis functions}
}
\value{
The corresponding matrix of basis functions
}
\description{
Make B matrix
}
\details{
Create set of basis functions using the Bfuncs list
}
