% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print values for a mars Object}
\usage{
\method{print}{mars}(mars)
}
\arguments{
\item{mars}{a mars object, usually created by mars}
}
\description{
Prints out a user-friendly output of the mars object details (includes the coefficients, basis functions, product and call)
}
\examples{
## Print output for fitting a mars to the mtcars data
mtmars = mars(mpg~., data=mtcars, control=mars.control())
print(mtmars)
}
\seealso{
Other methods: 
\code{\link{anova.mars}()},
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
