% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.mars.R
\name{print_product}
\alias{print_product}
\title{Print Product of mars object hinge functions}
\usage{
print_product(mars)
}
\arguments{
\item{mars}{a mars object, created using mars}
}
\description{
Allows the user to print the product of hinge functions for the passed in mars object
}
\details{
Print Product of mars Object hinge functions
}
\examples{
mtcars_mars = mars(mpg~., data=mtcars, control=mars.control());
print_product(mtcars_mars)
}
