% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{split_points}
\alias{split_points}
\title{Split Points}
\usage{
split_points(xv, Bm)
}
\arguments{
\item{xv}{Possible split points}

\item{Bm}{A basis function}
}
\value{
a sorted list of unique split points where the basis function is positive,
with the largest point removed
}
\description{
Find possible split points and sort them, removing the largest value
}
\references{
Lin, W. (2023). Lecture Material.
}
