% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary of a mars Object}
\usage{
\method{summary}{mars}(mars)
}
\arguments{
\item{mars}{a mars object, usually created by mars}
}
\value{
The function summary.mars computes and returns a list of summary
statistics of the fitted linear model given in object,
using the components (list elements) "call" and "terms" from its argument
}
\description{
Outputs a detailed summary of the mars model object (includes the coefficients, basis functions, residuals, additional statistics and other)
}
\examples{
m <-mars(Ozone~Wind+Temp+Month,data=airquality,control=mars.control())
summary(m)
}
\seealso{
Other methods: 
\code{\link{anova.mars}()},
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()}
}
\concept{methods}
