% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{validate_mars.control}
\alias{validate_mars.control}
\title{Validator for `mars.control` objects}
\usage{
validate_mars.control(control)
}
\arguments{
\item{control}{a mars.control object}
}
\value{
a mars.control object
}
\description{
Checks that a mars.control object has valid parameters.
}
\references{
Lin, W. (2023). Lecture Material.
}
