% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{bwd_stepwise}
\alias{bwd_stepwise}
\title{Backwards Stepwise}
\usage{
bwd_stepwise(fwd, control)
}
\arguments{
\item{fwd}{output created by running fwd_stepwise}

\item{control}{mars.control object, created with mars.control}
}
\value{
A list with elements
\describe{
\item{y The same vector of response values given as input}{}
\item{B }{}
\item{Bfuncs }{}
}
}
\description{
Implementation of Backwards Stepwise algorithm (algorithm
3 from page 17 of the Friedman paper)
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, 19(1),
 1–67. https://doi.org/10.1214/aos/1176347963
Lin, W. (2023). Lecture Material.
}
