% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Forward Stepwise}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{vector of response values}

\item{x}{dataset of predictor variables}

\item{control}{mars.control object created by mars.control}
}
\value{
A list with elements
\describe{
\item{y The same vector of response values given as input}{}
\item{B }{}
\item{Bfuncs }{}
}
}
\description{
Implementing algorithm 2 from the page 17 of the Friedman Paper
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, 19(1),
 1–67. https://doi.org/10.1214/aos/1176347963
}
