% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{Lack Of Fit}
\usage{
LOF(form, data, control)
}
\arguments{
\item{form}{formula to fit a linear model}

\item{data}{dataset to fit a model to}

\item{control}{a mars.control object created by mars.control}
}
\value{
lof, the lack of fit value for the linear model fit with the given data and formula
}
\description{
Use GCV to calculate the the lack of fit of a model, following equation 30
on page 20 of the Friedman paper
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, 19(1),
 1–67. https://doi.org/10.1214/aos/1176347963
}
