% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.mars.R
\name{print_basis}
\alias{print_basis}
\title{Print Basis Functions of mars object}
\usage{
print_basis(mars)
}
\arguments{
\item{mars}{a mars object, created using mars}
}
\description{
Allows the user to print the basis functions of the passed in mars object
}
\details{
Print Basis Functions of a mars Object
}
\examples{
mtcars_mars = mars(mpg~., data=mtcars, control=mars.control());
print_basis(mtcars_mars)
}
